# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CV Analyst API",
        "description": "\n * **CV Analyst API** - Solução completa para análise de CVs e gestão de oportunidades de emprego\n * \n * ## 🚀 Funcionalidades Principais:\n * - **Análise Automática de CVs**: Utilize IA para analisar e pontuar currículos\n * - **Gestão de Oportunidades**: Publique e gira vagas de emprego\n * - **Sistema Kanban**: Gestão visual do processo de recrutamento\n * - **API Pública**: Endpoints públicos para listagem de vagas\n * - **Relatórios**: Estatísticas e relatórios detalhados\n * \n * ## 🔐 Autenticação:\n * A API utiliza Laravel Sanctum com tokens Bearer. Primeiro registe-se ou faça login para obter um token, depois inclua-o no header: `Authorization: Bearer {seu-token}`\n * \n * ## 📝 Exemplos:\n * - Para testar endpoints protegidos, clique em 'Authorize' e insira seu token\n * - Endpoints públicos não requerem autenticação\n * - Todos os uploads de arquivos devem ser PDFs (máximo 5MB)\n * \n * ## 🌐 Ambientes:\n * - **Desenvolvimento**: {APP_URL}/api\n * - **Produção**: {PRODUCTION_API_URL}/api\n * \n * As URLs são configuradas dinamicamente através das variáveis de ambiente.\n * ",
        "contact": {
            "name": "Suporte CV Analyst",
            "url": "{PRODUCTION_FRONTEND_URL}",
            "email": "lucas.rodrigues@team.inovcorp.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "{PRODUCTION_FRONTEND_URL}/license"
...
    },
    "servers": [
...
    "paths": {
...
        "/public/empresa/{slug}": {
            "get": {
                "tags": [
                    "🌐 Público"
                ],
                "summary": "Visualizar página de empresa",
                "description": "Retorna informações da empresa e suas oportunidades públicas",
                "operationId": "677be344b1bdfa235be92b2850b00855",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug da empresa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "tech-solutions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações da empresa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "company": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Tech Solutions"
                                                },
                                                "website": {
                                                    "type": "string",
                                                    "example": "https://techsolutions.pt"
                                                },
                                                "sector": {
                                                    "type": "string",
                                                    "example": "Tecnologia"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "oportunidades": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Desenvolvedor Full Stack"
                                                    },
                                                    "descricao": {
                                                        "type": "string",
                                                        "example": "Procuramos um desenvolvedor..."
                                                    },
                                                    "skills_desejadas": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa não encontrada"
                    }
                }
            }
        },
...
        "/companies/login": {
            "post": {
                "tags": [
                    "🔐 Autenticação"
                ],
                "summary": "Login de empresa",
                "description": "Autentica uma empresa e retorna o token de acesso",
                "operationId": "ecc5183b811427a937d379f73981ee4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@techsolutions.pt"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login realizado com sucesso"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123def456..."
                                        },
                                        "company": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Tech Solutions Lda"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@techsolutions.pt"
                                                },
                                                "website": {
                                                    "type": "string",
                                                    "example": "https://techsolutions.pt"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciais inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Email não verificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email não verificado. Verifique o seu email para ativar a conta."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação Bearer obtido através do login. Formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}